fixed an issue where using this command in dms would crash the bot

+++ commands/clear.js
@@ -10,6 +10,7 @@ module.exports = {
     if (talkedRecently.has(message.author.id)) {
       message.channel.send("Please wait " + commandDelay + " second(s) until you can use this command again");
     } else {
+      if (message.guild == null) return message.reply("You cannot do this command in DMs.");
       if (!message.guild.members.cache.get(message.author.id).roles.cache.some(role => role.name === 'ModPerms')) return message.reply("You do not have the permission to do this command!")
       const amnt = parseInt(args[1]) + 1 || 2
       message.channel.bulkDelete(amnt)
